fix: remove unnecessary u64 cast
Remove redundant cast from u64 to u64 in block interval validation
to resolve clippy::unnecessary_cast warning.

+++ crates/multivm-cli/src/main.rs
@@ -159,7 +159,7 @@ async fn main() -> Result<(), Box<dyn std::error::Error>> {
 
     // Validate block interval for security
     let validated_interval =
-        validation::validate_block_interval(block_gen_config.block_interval_ms as u64)?;
+        validation::validate_block_interval(block_gen_config.block_interval_ms)?;
     let block_interval = validated_interval;
 
     let block_generator =
